InventoryManagement/Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
reportlab==4.0.4
